Pokedex_Backend/requirements.txt
Flask
flask_cors
pyodbc
azure-storage-blob